+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - reset the payloads of a stream at a stream restart
 - moved some code for a stream restart from `_keepalive_streams()` to `restart_stream()` which caused that the direct
 call of `restart_stream()` worked only inside of `_keepalive_streams()`
+- handling of `RuntimeWarning` in class connection at row 189
 
 ## 1.13.0
 ### Added